Feat: 카테고리 선택하는 단계 컴포넌트 분리
